+++ src/driver/graphic.rs
@@ -5,8 +5,11 @@ use {
         device::Device,
         image::SampleCount,
         merge_push_constant_ranges,
-        shader::{DescriptorBindingMap, PipelineDescriptorInfo, Shader, SpecializationInfo},
-        DriverError,
+        shader::{
+            DescriptorBindingMap, DescriptorInfo, PipelineDescriptorInfo, Shader,
+            SpecializationInfo,
+        },
+        DescriptorBinding, DriverError,
     },
     ash::vk,
     derive_builder::{Builder, UninitializedFieldError},
@@ -363,6 +366,7 @@ pub struct GraphicPipeline {
     pub name: Option<String>,
 
     pub(crate) push_constants: Vec<vk::PushConstantRange>,
+    pub(crate) separate_samplers: Box<[DescriptorBinding]>,
     pub(crate) shader_modules: Vec<vk::ShaderModule>,
     pub(super) state: GraphicPipelineState,
 }
@@ -458,6 +462,13 @@ impl GraphicPipeline {
                 descriptor_info.set_binding_count(info.bindless_descriptor_count);
             }
         }
+
+        let separate_samplers = descriptor_bindings
+            .iter()
+            .filter_map(|(&descriptor_binding, (descriptor_info, _))| {
+                matches!(descriptor_info, DescriptorInfo::Sampler(..)).then_some(descriptor_binding)
+            })
+            .collect();
 
         let descriptor_info = PipelineDescriptorInfo::create(&device, &descriptor_bindings)?;
         let descriptor_sets_layouts = descriptor_info
@@ -560,6 +571,7 @@ impl GraphicPipeline {
                 layout,
                 name: None,
                 push_constants,
+                separate_samplers,
                 shader_modules,
                 state: GraphicPipelineState {
                     layout,